bug in codec function

+++ src/codecs/text_md_codec.h
@@ -25,7 +25,7 @@
 class TextMDCodec : public AbstractMDCodec{
 private:
 	Uint8 m_flows_number;
-	Uint8 m_descriptors_number;
+	Uint32 m_descriptors_number;
 	Uint16 m_descr_dim_total;
 public:
